menu di pausa e skill
Aggiunto sistema di PA recovery quando si mette a segno un colpo base
Implementato un sistema di lettura da un file contenente le info su tutte le skill e aggiunta per ogni personaggio delle skill da quella lista
Aggiunta nome, sprite e moltiplicatore attacco per arma, aggiunto anche monete e icone abilità nel menù di pausa (per ora il menù di pausa visualizza le informazioni ma non permette di cambiarle)

Ricordarsi di aggiungere un lettura delle abilità apprese dal giocatore scrivendole e leggendole da apposito file

+++ Assets/Script/Damage.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 public class Damage
 {
     public Vector3 origin;
-    public int attackPower;     //Il danno è calcolato come AttaccoBase*ATKAttaccante/DEFRicevennte. attackPower rappresenta il calcolo parziale AttaccoBase*ATKAttaccante 
+    public float attackPower;     //Il danno è calcolato come AttaccoBase*ATKAttaccante/DEFRicevennte. attackPower rappresenta il calcolo parziale AttaccoBase*ATKAttaccante 
     public float pushForce;
 }
